@@ -23,15 +23,19 @@
   },
   "dependencies": {},
   "devDependencies": {
-    "tsd": "^0.6.0-beta.5",
-    "typescript": "1.5.0-alpha"
+    "tsconfig-glob": "^0.4.1",
+    "tsconfig-lint": "^0.5.0",
+    "typescript": "^1.7.5",
+    "typings": "^0.6.7"
   },
   "bugs": {
     "url": "https://github.com/Platypi/ts-bundle/issues"
   },
   "scripts": {
-    "prepublish": "npm run tsd && npm run build",
-    "tsd": "./node_modules/.bin/tsd update -so",
-    "build": "./node_modules/.bin/tsc -p ."
+    "prepublish": "npm run typings && npm run build",
+    "typings": "typings install",
+    "build": "tsconfig && tsc -p .",
+    "tsconfig": "tsconfig",
+    "lint": "tsconfig-lint --use-glob"
   }
 }
