@@ -1,3 +1,2 @@
-﻿/// <reference path="references.d.ts" />
-import compress = require('./lib/compressor');
+﻿import compress = require('./lib/compressor');
 export = compress;
