@@ -1,26 +1,25 @@
-﻿/// <reference path="../references.d.ts" />
+﻿import Module from './module';
+import Writer from './writer';
+import buildModule from './buildmodule';
+import * as globals from './globals';
 
-import Module = require('./module');
-import Writer = require('./writer');
-import buildModule = require('./buildmodule');
-import globals = require('./globals');
-
-var commentsRemainingRegex = globals.commentsRemainingRegex,
+let commentsRemainingRegex = globals.commentsRemainingRegex,
     whitespaceRegex = globals.whitespaceRegex,
     intermediarySpaceRegex = globals.intermediarySpaceRegex;
 
-function buildContents(lines: Array<string>) {
-    var root = globals.rootModule,
+export default function buildContents(lines: Array<string>) {
+    let root = globals.rootModule,
         prepend = root.name !== globals.windowName ? '    ' : '',
         moduleName = '',
         previousLine = '',
         currentLines: Array<string> = [],
         commentLines: Array<string> = [],
-        line: string;
+        line: string,
+        index = 0;
 
-    for (var index = 0; index < lines.length; ++index) {
+    for (; index < lines.length; ++index) {
         line = globals.removeCommentsAndStrings(lines[index]);
-        
+
         // Handles comment blocks
         while (line.indexOf('/*') !== -1) {
             if (index > 0 && !!previousLine) {
@@ -46,10 +45,10 @@ function buildContents(lines: Array<string>) {
                 line = lines[index];
             }
 
-            var substringLine = line.substr(line.indexOf('*/') + 2);
+            let substringLine = line.substr(line.indexOf('*/') + 2);
             substringLine = globals.removeCommentsAndStrings(substringLine);
 
-            // Check if there is another comment block on the same line as the 
+            // Check if there is another comment block on the same line as the
             // end comment.
             if (substringLine.indexOf('/*') === -1) {
                 commentLines.push(prepend + line);
@@ -58,15 +57,15 @@ function buildContents(lines: Array<string>) {
             } else if (substringLine.indexOf('/*') === 0) {
                 line = substringLine;
             } else {
-                // We have a new comment block, so we need to cut out the whitespace 
-                // between the end comment and beginning comment, and add a new line 
+                // We have a new comment block, so we need to cut out the whitespace
+                // between the end comment and beginning comment, and add a new line
                 // to break up the comment block.
-                var i = lines[index].indexOf('/*'),
+                let i = lines[index].indexOf('/*'),
                     whitespace = lines[index].match(whitespaceRegex)[0];
 
                 lines.splice(index + 1, 0, whitespace + lines[index].substr(i));
 
-                var startSubstring = lines[index].indexOf('*/') + 2;
+                let startSubstring = lines[index].indexOf('*/') + 2;
                 lines.splice(index + 1, 0, whitespace + lines[index].substr(startSubstring, i - startSubstring).trim());
                 lines[index] = lines[index].substring(0, startSubstring);
 
@@ -81,7 +80,7 @@ function buildContents(lines: Array<string>) {
         line = line.replace(commentsRemainingRegex, '').trim();
 
         if (line.indexOf('module ') === 0) {
-            // We found a new module, so we need to add the remaining currentLines 
+            // We found a new module, so we need to add the remaining currentLines
             // to the root module, then build the new module lines.
             if (currentLines.length > 0) {
                 root.writers.push(new Writer(currentLines));
@@ -89,7 +88,7 @@ function buildContents(lines: Array<string>) {
             }
             moduleName = line.replace('{', '').split(intermediarySpaceRegex)[1];
 
-            var mod = Module.fetch(root, moduleName);
+            let mod = Module.fetch(root, moduleName);
             mod.docs.push(new Writer(commentLines));
             index = buildModule(mod, lines, index + 1);
 
@@ -111,5 +110,3 @@ function buildContents(lines: Array<string>) {
     root.writers.push(new Writer(currentLines));
     return index;
 }
-
-export = buildContents;
