@@ -1,23 +1,21 @@
-﻿/// <reference path="../references.d.ts" />
-
-import globals = require('./globals');
+﻿import * as globals from './globals';
 
 /**
- * Iterates through the index.html file data, finding all the 
- * script tags between the start and end comment nodes. Takes 
- * the src value of each script tag and changes the extension 
- * from .js to .ts. Returns an array of all the .ts files in 
+ * Iterates through the index.html file data, finding all the
+ * script tags between the start and end comment nodes. Takes
+ * the src value of each script tag and changes the extension
+ * from .js to .ts. Returns an array of all the .ts files in
  * the same order they appear in the index.html.
- * 
- * @param data The string output from reading the index.html 
+ *
+ * @param data The string output from reading the index.html
  * file.
  */
-function getFiles(data: string) {
+export default function getFiles(data: string) {
     if (typeof data !== 'string') {
         return [];
     }
 
-    var lines = data.split(/\r\n|\n/),
+    let lines = data.split(/\r\n|\n/),
         files: Array<string> = [],
         running = false;
 
@@ -30,7 +28,7 @@ function getFiles(data: string) {
             running = true;
             return false;
 
-        // If we hit the end tag, we need to stop checking for script tags. 
+        // If we hit the end tag, we need to stop checking for script tags.
         } else if (running && globals.endRegex.test(line)) {
             return true;
         // We want to ignore script tags if we haven't hit a start tag yet.
@@ -38,7 +36,7 @@ function getFiles(data: string) {
             return false;
         }
 
-        var exec = globals.fileRegex.exec(line);
+        let exec = globals.fileRegex.exec(line);
 
         // Ignore the tag if it is not a script tag with a *.js src value.
         if (!exec) {
@@ -52,5 +50,3 @@ function getFiles(data: string) {
 
     return files;
 }
-
-export = getFiles;
