@@ -1,21 +1,19 @@
-﻿/// <reference path="../references.d.ts" />
+﻿import Module from './module';
 
-import Module = require('./module');
-
-class Writer {
+export default class Writer {
     private __module: Module;
     private __lines: Array<string>;
 
     /**
      * Creates a new Writer with the given lines.
-     * 
+     *
      * @param lines The lines to set for the Writer.
      */
     constructor(lines: Array<string>);
 
     /**
      * Creates a new Writer with the given module.
-     * 
+     *
      * @param mod The module to set for this writer.
      */
     constructor(mod: Module);
@@ -28,16 +26,16 @@ class Writer {
     }
 
     /**
-     * Writes the lines for this writer into the output array, then calls the callback with the 
+     * Writes the lines for this writer into the output array, then calls the callback with the
      * module associated with this writer.
-     * 
+     *
      * @param output The output lines to which the Writer's lines should be appended.
      * @param previousLine The previous line that was added to output.
-     * @param callback A method that takes in the next module with which to continue 
+     * @param callback A method that takes in the next module with which to continue
      * output generation.
      */
     write(output: Array<string>, previousLine: string, callback: (mod: Module) => void) {
-        // If this writer has a module associated with it, we want to first add the 
+        // If this writer has a module associated with it, we want to first add the
         // lines for the module before adding the lines for this writer.
         if (!!this.__module) {
             if (!!callback) {
@@ -49,7 +47,7 @@ class Writer {
             return previousLine;
         }
 
-        var trim: string;
+        let trim: string;
 
         // Add every line to the output.
         this.__lines.forEach((line) => {
@@ -65,5 +63,3 @@ class Writer {
         return previousLine;
     }
 }
-
-export = Writer;
